firewall zone support with firewalld
The firewall role has been extended to be able to support the definition
of the zone a change is ensured on. A given zone will be used for these
parameters: `service`, `port` and `forward_port` without a given interface
or MAC address.

Also minor code cleanup has been done to enable the use of zones.