fix: Skip runtime operations in non-systemd environments
Cause: Both the actual role and the tests were previously assuming that
the system was booted with systemd and could run services. But this is
not the case when running the role during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not.

Result: The role now works during container builds.

Call our firewall modules with `online: false` to select their
`firewall-offline-cmd` modes when not booted.

Add a new "online" parameter to get_files_checksums.sh that skips the
reload when offline.

Adjust tests_firewall_fact.yml to ensure that "detailed" mode fact
collection fails in container mode (as that is currently unsupported,
see previous commit).

https://issues.redhat.com/browse/RHEL-88425