ci: Assert fact structure and some well-known entries
Spot-check a particular entry of each fact group (helpers, icmptypes,
services, and zones) to ensure the returned fact structure is as
expected. This will help with validating a future `firewall-offline-cmd`
backend. Note that "policies" does not yet exist in RHEL 7/8, so don't
check that.

Also check that the service addition and custom service creation actions
are actually reflected in the facts.