feat: Add offline mode to firewall_lib_facts
When calling this role with `offline: false`, collect the facts using
`firewall-offline-cmd`, which will work during container builds.

Don't implement `detailed` mode yet -- there isn't much use of that
during container builds, as there should be zero surprise about the
current state. It is possible to implement if we ever need it, but let's
only spend that time if/when we need it.