fix: Fix "helpers" service option
Cause: The query check was inverted, and query/add used the wrong API (module,
not helper).

Consequence: Trying to set the "helpers:" option of a created service
had no effect.

Fix: Use the correct API and fix the check. Add an integration test
which actually validates the created service in firewalld.

The service helper API does not yet exist on RHEL 7, so skip the
corresponding tests on RHEL < 8.

Fixes #276